requirements.txt
ops >= 1.4.0

MarkupSafe
pyyaml
ansible
